requirements.txt
# -*- coding: utf-8 -*-
python-telegram-bot==13.7
requests==2.31.0
python-dotenv==1.0.0